Change the leaderboard from randomly sorted to sorted in order from highest score to the lowest score. To change a list to a array you push a object into the list.
